+++ lib/rbscmlex/token.rb
@@ -36,26 +36,26 @@ module Rbscmlex
   # a structure to store properties of a token of Scheme program.
 
   Token = Struct.new(:type, :literal) {
+    # :stopdoc:
+    # `to_a` and `to_h` are automatically defined for a class
+    # generated from Struct.
+    # :startdoc:
+
     alias :to_s :literal
+
+    # Generates a new string of JSON notation, which has "type" and
+    # "literal" as its key.
+    def to_json
+      JSON.generate(to_h)
+    end
   }
 
   class << self
 
-    # Instantiates a new token object form type and literal.  The 3rd
-    # argument specifies the form of the object.  It must be one of
-    # :token, :hash, and :json.
+    # Instantiates a new token object form type and literal.
 
-    def new_token(type, literal = nil, form = :token)
-      case form
-      when :token
-        Token.new(type, literal)
-      when :hash
-        make_hash(type, literal)
-      when :json
-        JSON.generate(make_hash(type, literal))
-      else
-        raise InvalidConversionTypeError, "cannot generate #{type} as token"
-      end
+    def new_token(type, literal = nil)
+      Token.new(type, literal)
     end
 
     # Returns true when the argument is valid token type.
@@ -64,7 +64,7 @@ module Rbscmlex
       TOKEN_TYPES.include?(type)
     end
 
-    # Returns a new Hash object with type and literal.
+    # Returns a new Hash object with type and literal as its keys.
 
     def make_hash(type, literal)
       {type: type, literal: literal}
@@ -75,28 +75,54 @@ module Rbscmlex
     # must be valid token type.  Otherwise, raises UnknownTokenTypeError.
 
     def hash2token(hash)
-      if h.key?("type") and h.key?("literal")
-        type = h["type"].intern
+      if hash.key?(:type) and hash.key?(:literal)
+        type = hash[:type].intern
         raise UnknownTokenTypeError, ("got=%s" % type) unless token_type?(type)
-        literal = h["literal"]
-        Token.new(type.intern, literal)
+        literal = hash[:literal]
+        new_token(type, literal)
       else
         raise InvalidHashError, ("got=%s" % hash)
       end
     end
 
     # Converts a JSON notation, which hash type and literal, to a new
-    # token object.  The value associated to type of the Hash must be
+    # token object.  The value associated to type of the JSON must be
     # valid token type.  Otherwise, raises UnknownTokenTypeError.
 
     def json2token(json)
-      h = JSON.parse(json)
+      h = JSON.parse(json, symbolize_names: true)
       begin
         hash2token(h)
       rescue InvalidHashError => _
         raise InvalidJsonError, ("got=%s" % json)
       end
     end
+
+    # Converts a Token object to a Hash object.
+
+    def token2hash(token)
+      token.to_h
+    end
+
+    # Converts a Token object to a string of JSON notation.
+
+    def token2json(token)
+      token.to_json
+    end
+
+    # Converts a Hash object to a string of JSON notation.
+
+    def hash2json(hash)
+      JSON.generate(hash)
+    end
+
+
+    # Converts a JSON notation to a new Hash object.
+
+    def json2hash(json)
+      JSON.parse(json)
+    end
+
   end
 
 end