+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 - (nothing to record here)
 
+## [0.1.2] - 2021-05-07
+### Added
+- Add a mechanism to initialize a Parser instance from an array of
+  tokens, which already created from source of Scheme.
+
+### Fixed
+- Fix issue #3: a wrong link in `README.md`.
+
 ## [0.1.1] - 2021-05-06
 ### Fixed
 - Fix issue #1: `rbscmlex` fails to read from STDIN.